Update index.js
Permit use of the process.env variable via Vitejs if the user don't use Cra.
ViteJs need a variable starting with VITE_ to be able to use it. 
Even if we normally need to use import.meta.env with vite, the developper can do the necessary job to react here.
But the correct process.env varable need to be on reach on json-graphql-parser.